🏗️ Architect: Modularized Place
- Refactored the 744-line `apts/place.py` 'God Object' into a modular package `apts/place/`.
- Isolated core `Place` logic and state in `base.py`.
- Extracted `TFProxy`, `get_scalar_datetime`, and cached search functions to `utils.py`.
- Decomposed `Place` functionality into four specialized MixIns in `models.py`:
  - `PlaceConditionsMixIn`: SQM, seeing, and light pollution calculations.
  - `PlaceTimesMixIn`: Rise, set, and twilight event timing.
  - `PlacePathsMixIn`: Sun/Moon altitude/azimuth curves and path plotting.
  - `PlaceImagingMixIn`: Imaging window determination.
- Maintained the public API and backward compatibility via re-exports in `__init__.py`.
- Verified the refactor with the full unit test suite (663/663 tests passed).
